Introduce INodeDoc instead of passing around dicts
INodeDoc's are returned when you query the registry, and are read-only
snapshots of the nodes in the registry.  It also serves to document the
attributes supported by the registries.

INodeDoc is implemented by registry.nodedoc.NodeDoc which is used by
registry.couch_registry.CouchRegistry and by
registry.dict_registry.DictRegistry.